+++ gallery/admin_urls.py
@@ -1,0 +1,14 @@
+from django.conf.urls.defaults import patterns, url
+
+
+urlpatterns = patterns(
+    '',
+
+    url(
+        r'^gallery/gallery/(?P<gallery_id>\d+)/bulk-image-upload/$',
+        'gallery.admin_views.bulk_image_upload',
+        {},
+        name='gallery-bulk-image-upload'
+    )
+
+)