+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+next
+----
+#. Bulk image upload form
+
 0.2
 ---
 #. Refactor embedded video so it can be used in a gallery.